start template coverage test
