+++ src/templates/index.js
@@ -1,61 +1,11 @@
 const Infobox = require('../infobox/Infobox');
 const Reference = require('../reference/Reference');
-const getName = require('./_parsers/_getName');
 const getTemplates = require('./_parsers/_getTemplates');
-
-const generic = require('./generic');
-const ignore = require('./_ignore');
+const parseTemplate = require('./parse');
 
 //ensure references and infoboxes at least look valid
 const isObject = function(x) {
   return (typeof x === 'object') && (x !== null) && x.constructor.toString().indexOf('Array') === -1;
-};
-
-//put them all together
-const bigParsers = require('./01-data');
-const inlineParsers = require('./02-text');
-
-//this gets all the {{template}} strings and decides how to parse them
-const oneTemplate = function(tmpl, wiki, data, options) {
-  let name = getName(tmpl);
-
-  //we explicitly ignore these templates
-  if (ignore.hasOwnProperty(name) === true) {
-    wiki = wiki.replace(tmpl, '');
-    return wiki;
-  }
-
-  //string-replacement templates
-  if (inlineParsers.hasOwnProperty(name) === true) {
-    let str = inlineParsers[name](tmpl, data);
-    wiki = wiki.replace(tmpl, str);
-    return wiki;
-  }
-
-  //section-template parsers
-  if (bigParsers.hasOwnProperty(name) === true) {
-    let obj = bigParsers[name](tmpl);
-    if (obj) {
-      data.templates.push(obj);
-    }
-    wiki = wiki.replace(tmpl, '');
-    return wiki;
-  }
-
-  //fallback parser
-  let obj = generic(tmpl, name);
-  if (obj) {
-    data.templates.push(obj);
-    wiki = wiki.replace(tmpl, '');
-    return wiki;
-  }
-  //bury this template, if we don't know it
-  if (options.missing_templates === true) {
-    console.log(':: ' + name);
-  }
-  wiki = wiki.replace(tmpl, '');
-
-  return wiki;
 };
 
 //reduce the scary recursive situations
@@ -63,7 +13,7 @@ const parseTemplates = function(wiki, data, options) {
   let templates = getTemplates(wiki);
   //first, do the nested (second level) ones
   templates.nested.forEach(tmpl => {
-    wiki = oneTemplate(tmpl, wiki, data, options);
+    wiki = parseTemplate(tmpl, wiki, data, options);
   });
   //then, reparse wiki for the top-level ones
   templates = getTemplates(wiki);
@@ -71,13 +21,13 @@ const parseTemplates = function(wiki, data, options) {
   //okay if we have a 3-level-deep template, do it again (but no further)
   if (templates.nested.length > 0) {
     templates.nested.forEach(tmpl => {
-      wiki = oneTemplate(tmpl, wiki, data, options);
+      wiki = parseTemplate(tmpl, wiki, data, options);
     });
     templates = getTemplates(wiki); //this is getting crazy.
   }
   //okay, top-level
   templates.top.forEach(tmpl => {
-    wiki = oneTemplate(tmpl, wiki, data, options);
+    wiki = parseTemplate(tmpl, wiki, data, options);
   });
   //lastly, move citations + infoboxes out of our templates list
   let clean = [];
